third_party/blink/web_tests/platform/generic/fast/dom/Element/offsetLeft-offsetTop-body-quirk-expected.txt

static: body: (0, 0) child: (27, 27)
relative: body: (0, 0) child: (17, 17)
fixed: body: (0, 0) child: (17, 17)
absolute: body: (0, 0) child: (17, 17)
